Fix misc. compiler warnings.
Found with gcc (GCC) 8.4.1 20200928 (Red Hat 8.4.1-1).

- cmd_line.c:579: warning: ‘strncat’ output may be truncated copying 151
                   bytes from a string of length 151 [-Wstringop-truncation]
- hwloc_base_util.c:1050: warning: ‘strncat’ output may be truncated
                           copying between 0 and 8191 bytes from a string
                               of length 8191 [-Wstringop-truncation]
- base/hwloc_base_util.c:1077: warning: ‘strncat’ output may be truncated
                                copying between 0 and 8191 bytes from a string
                                of length 8191 [-Wstringop-truncation]
- base/mpool_base_alloc.c:81: warning: comparison of integer expressions of
                               different signedness: ‘ssize_t’ {aka ‘long int’}
                               and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
- mpiext_affinity_str.c:200: warning: ‘strncat’ output may be truncated copying between
                              0 and 1024 bytes from a string of length 8191
                              [-Wstringop-truncation]
- mpiext_affinity_str.c:235: warning: ‘strncat’ output may be truncated copying between
                              0 and 1024 bytes from a string of length 8191
                              [-Wstringop-truncation]
- mpiext_affinity_str.c:251: warning: ‘strncat’ output may be truncated copying between
                              0 and 1024 bytes from a string of length 8191
                              [-Wstringop-truncation]
- mpiext_affinity_str.c:273: warning: ‘strncat’ output may be truncated copying between
                              0 and 1024 bytes from a string of length 8191
                              [-Wstringop-truncation]
- mpiext_affinity_str.c:226: warning: ‘strncat’ output may be truncated copying between
                              0 and 1024 bytes from a string of length 8191
                              [-Wstringop-truncation]
- hook_comm_method_fns.c:32: warning: ‘strncpy’ output may be truncated copying 15 bytes
                              from a string of length 63 [-Wstringop-truncation]
- osc_sm_component.c:224:32: warning: comparison of integer expressions of different
                              signedness: ‘long unsigned int’ and ‘ssize_t’
                              {aka ‘long int’} [-Wsign-compare]
- osc_sm_component.c:262:16: warning: declaration of ‘flag’ shadows a previous local
                              [-Wshadow]
- osc_sm_component.c:413:13: warning: declaration of ‘flag’ shadows a previous local
                              [-Wshadow]
- osc_rdma_component.c:1371: warning: comparison of integer expressions of different
                              signedness: ‘long unsigned int’ and ‘ssize_t’
                              {aka ‘long int’} [-Wsign-compare]
- osc_rdma_component.c:1325: warning: unused variable ‘infoval’ [-Wunused-variable]
- dpm/dpm.c:667: warning: ‘modifiers’ may be used uninitialized in this function
                  [-Wmaybe-uninitialized]

- Cleanup: Move common code to osc/base.

(cherry picked from commit d4d5ff3ea727d93020c28fbef21333b1601475a9)

+++ opal/mca/mpool/base/mpool_base_alloc.c
@@ -63,7 +63,7 @@ void *mca_mpool_base_alloc(size_t size, opal_info_t *info, const char *hints)
     mca_mpool_base_module_t *mpool;
     void *mem = NULL;
     opal_cstring_t *align_info_str;
-    size_t memory_alignment = OPAL_ALIGN_MIN;
+    long long memory_alignment = OPAL_ALIGN_MIN;
 
     mpool_tree_item = mca_mpool_base_tree_item_get();
     if (!mpool_tree_item) {
@@ -76,7 +76,7 @@ void *mca_mpool_base_alloc(size_t size, opal_info_t *info, const char *hints)
                       &align_info_str, &flag);
 
         if (flag) {
-            ssize_t tmp_align = atoll(align_info_str->string);
+            long long tmp_align = atoll(align_info_str->string);
             OBJ_RELEASE(align_info_str);
             if (tmp_align > memory_alignment) {
                 memory_alignment = tmp_align;
